sdk/python/pyproject.toml
[project]
name = "uos-kernel-sdk"
version = "0.1.0"
description = "Python SDK for UOS Kernel gRPC API"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["uos_kernel*", "gen*"]
